Move styles to another file. Right align name
